rcar-vin: Replace group lock with graph mutex
The group mutex lock is to protect the group mask, vin, and csi
members. These essentially represent the state of a portion of
the media graph, and there is already a mutex to protect changes
to the media graph, so it is more correct to use the latter lock
for this purpose. Replace the group lock with the media_device
graph_mutex.

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>