rcar-vin: Move media_device_register to async completion
The .link_notify callback checks if the master VIN device is present and
returns -ENODEV if not. This is only necessary because the media_device
is registered before async completion, so it is possible for .link_notify
to be called before all devices are bound.

Fix this by moving media_device_register() to rvin_group_notify_complete().
This ensures that all devices are now bound (the rcar-csi2 subdevices and
and video capture devices) before .link_notify can be called.

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>