rcar-vin: Attach video device lists to subdevice source pads
Create a list of video capture devices that can be reached from a
source pad, and attach this list to every subdevice source pad in
the media graph.

The lists are composed of an array of list_heads attached to the
subdevice's host_priv pointer, one list_lead for every pad. Each
list_head in the array is the list of video devices reachable in
the graph from a pad.

The vdev lists can be used to inherit controls from subdevices, and to
forward events from subdevices, to every reachable video capture device
in the current graph.

The vdev lists for every subdevice source pad in the media graph have to be
updated whenever any link is created or removed. So rvin_group_link_notify()
catches the MEDIA_DEV_NOTIFY_LINK_CREATED and MEDIA_DEV_NOTIFY_LINK_REMOVED
notifications and either adds new connected video devices to source pads
when a link is created, or prunes video devices from source pads if
they are no longer connected when a link is removed. The vdev list update
is carried out by the function rvin_group_update_vdev_lists().

The 'path_enabled' boolean in a vdev list entry also needs to be updated
whenever a link in the graph is enabled or disabled. So
rvin_group_link_notify() calls rvin_group_update_vdev_enabled_paths()
on a MEDIA_DEV_NOTIFY_POST_LINK_CH notification.

Signed-off-by: Steve Longerbeam <steve_longerbeam@mentor.com>